requirements.txt
gspread
oauth2client
openai
requests
python-dotenv
tweepy
pytz